# HolidaysApp-Backend/driving/api-rest/contracts/holidayapp-openapi-v3.yaml
openapi: 3.0.0
info: # Información general sobre la API
  title: HolidaysApp API Rest
  version: 1.0.0
  description: API Rest for HolidaysApp
servers: # URLs base del servidor
  - url: http://localhost:8080
    description: Generated server url
...
paths: # Todos los endpoints de la API
...
  /api-rest/users:
    get:
      tags:
        - Users
      summary: Get all Users (optionally filtered by name or last name)
      description: Get all users or filter by name or last name using query parameter
      operationId: getAllUsers
      parameters:
        - name: username
          in: query
          description: Optional username to filter users
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid input data
        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The User was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
  /api-rest/users/login:
    post:
      tags:
        - Users
      summary: Login with the User's username or email and password
      description: Login with the User's username and password using a request body
      operationId: getUserByUsernameOrEmailAndPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid input data
        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The User was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
  /api-rest/users/register:
    post:
      tags:
        - Users
      summary: Register with the User's username, password and repeated password
      description: Login with the User's username and password using a request body
      operationId: registerByUsernameEmailAndPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid input data
        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The User was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
...
components:
  schemas:
...
    RoleDTO:
      description: This is the Role model
      type: object
      properties:
        id:
          type: integer
          #format: int32
        name:
          type: string
...
    UserDTO:
      description: This is the User model
      type: object
      properties:
        id:
          type: integer
        role:
          $ref: '#/components/schemas/RoleDTO'
        username:
          type: string
        email:
          type: string
        password:
          type: string
        enabled:
          type: boolean
    LoginRequest:
      description: This is the Login Request model
      type: object
      required:
        - usernameOrEmail
        - password
      properties:
        usernameOrEmail:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 8
          maxLength: 30
          pattern: '^(?=.*[A-Z])(?=.*[^a-zA-Z0-9]).*$'
          description: >
            Must contain at least one uppercase letter and one special character
    RegisterRequest:
      description: This is the Register Request model
      type: object
      required:
        - username
        - email
        - password
        - repeatPassword
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
          minLength: 8
          maxLength: 30
        password:
          type: string
          minLength: 8
          maxLength: 30
          pattern: '^(?=.*[A-Z])(?=.*[^a-zA-Z0-9]).*$'
          description: >
            Must contain at least one uppercase letter and one special character
        repeatPassword:
          type: string
          minLength: 8
          maxLength: 30
          pattern: '^(?=.*[A-Z])(?=.*[^a-zA-Z0-9]).*$'
          description: >
            Must contain at least one uppercase letter and one special character
...
    ErrorResponse:
      description: Standard error model for API responses
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        message:
          type: string